azure_function_app/requirements.txt
azure-functions
azure-functions-durable
azure-keyvault-secrets==4.2.0
azure-identity==1.5.0
azure-storage-file-datalake==12.2.1
pandas==1.2.3
msal==1.9.0
requests==2.21.0
csv-diff==1.1